pyproject.toml
[project]
name = "datalab-oss-container"
version = "0.11.0"
description = "Containerized inference service for `marker`"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[project.urls]
Homepage = "https://datalab.to"
Repository = "https://github.com/datalab-to/oss_container"

[dependency-groups]
dev = [
    "aiohttp",
    "aiofiles",
    "docker",
    "pypdfium2",
    "click",
    "ruff",
    "pre-commit>=4.2.0",
    "psutil",
    "nvidia-ml-py3",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "httpx>=0.28.1",
]
server = [
    "aio-pika>=9.5.5",
    "bs4>=0.0.2",
    "fastapi>=0.115.12",
    "pypdfium2>=4.30.0",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.34.0",
]
worker = [
    "marker-pdf==1.9.3",
    "surya-ocr==0.16.7",
    "pika>=1.3.2",
    "torch==2.8.*",
]
